ENH: replace parameters dict
This commit gets rid of the defaults.conf file and the 'parameters' dict
that was used to handle the configuration of bots. It instead uses class
variables and the defaults are set in the relevant parts of the code,
but can be easily overriden using environment variables or configuration
files.